agents/crewai/runner.py
import time
from typing import Dict, Any
from agents.common.agent_interface import AgentRunner
from agents.common.rag_client import RAGClient

class CrewAIRunner(AgentRunner):
    """
    Implementation of the AgentRunner interface for CrewAI.
    """
    
    def __init__(self, rag_service_url: str):
        """
        Initialize the CrewAI runner.
        
        Args:
            rag_service_url: URL of the RAG service
        """
        self.rag_client = RAGClient(rag_service_url)
        
    def run(self, task: str) -> Dict[str, Any]:
        """
        Run the CrewAI agents on a given task.
        
        Args:
            task: The task description
            
        Returns:
            Dictionary containing the agent's results and metrics
        """
        start_time = time.time()
        
        rag_queries = [
            {
                "query": f"Information needed for: {task}",
                "retrieved_context": "Placeholder context from RAG service",
                "usage": "Used to understand the task requirements"
            }
        ]
        
        time.sleep(1)  # Simulate processing time
        
        end_time = time.time()
        
        return {
            "result": f"CrewAI solution for: {task}",
            "reasoning": "Step-by-step reasoning process would go here",
            "rag_queries": rag_queries,
            "metrics": {
                "time_taken": end_time - start_time,
                "tokens_used": 1500,
                "rag_calls": len(rag_queries)
            }
        }

if __name__ == "__main__":
    import os
    from dotenv import load_dotenv
    
    load_dotenv()
    
    rag_service_url = f"http://{os.getenv('RAG_SERVICE_HOST', 'localhost')}:{os.getenv('RAG_SERVICE_PORT', '8000')}"
    runner = CrewAIRunner(rag_service_url)
    
    result = runner.run("Summarize the key points about climate change")
    print(result)
