requirements.txt
# Core Dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0

# RAG Dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.0.267
langchain-openai>=0.0.2

# Agent Framework Dependencies
crewai>=0.1.0
autogen>=0.1.0
langgraph>=0.0.10

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Utilities
numpy>=1.24.3
pandas>=2.0.0
tqdm>=4.65.0
